refactor: split settings view
